// src/materials/enums/material-type.ts
export enum MaterialType {
    MICKEY_EMBLEM = 'mickey-emblem',
    MINNIE_EMBLEM = 'minnie-emblem',
    DONALD_EMBLEM = 'donald-emblem',
    JOY_EMBLEM = 'joy-emblem',
    SADNESS_EMBLEM = 'sadness-emblem',
    ANXIETY_EMBLEM = 'anxiety-emblem',
    VANELLOPE_EMBLEM = 'vanellope-emblem',
    RALPH_EMBLEM = 'ralph-emblem',
    CALHOUN_EMBLEM = 'calhoun-emblem',
    FLOWER = 'flower',
}